refactor: move packages to internal
