+++ cmd/aio/main.go
@@ -10,10 +10,10 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 
-	"github.com/hetznercloud/csi-driver/api"
-	"github.com/hetznercloud/csi-driver/app"
-	"github.com/hetznercloud/csi-driver/driver"
-	"github.com/hetznercloud/csi-driver/volumes"
+	"github.com/hetznercloud/csi-driver/internal/api"
+	"github.com/hetznercloud/csi-driver/internal/app"
+	"github.com/hetznercloud/csi-driver/internal/driver"
+	"github.com/hetznercloud/csi-driver/internal/volumes"
 	"github.com/hetznercloud/hcloud-go/v2/hcloud/metadata"
 )
