+++ internal/api/helper.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	"github.com/hetznercloud/csi-driver/csi"
+	"github.com/hetznercloud/csi-driver/internal/csi"
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
