+++ test/integration/cryptsetup_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/go-kit/log"
 
-	"github.com/hetznercloud/csi-driver/volumes"
+	"github.com/hetznercloud/csi-driver/internal/volumes"
 )
 
 func TestCryptSetup(t *testing.T) {
